no errors shown in the program as the header files and new c files are getting bridged
