+++ myUpscaler/up60p_cli.h
@@ -1,0 +1,8 @@
+#ifndef UP60P_CLI_H
+#define UP60P_CLI_H
+
+int interactive_mode(const char *self_path);
+
+int process_cli_args(int argc, char **argv, const char *ffmpeg_path);
+
+#endif
