+++ myUpscaler/up60p_settings.h
@@ -5,7 +5,6 @@
 
 struct Settings {
     
-    
     char codec[8]; char crf[16]; char preset[32];
     char fps[16];
     char scale_factor[16];
@@ -72,6 +71,21 @@ struct Settings {
     char hwaccel[16]; char encoder[16];
 };
 
+void init_paths(void);
+void set_defaults(void);
+void reset_to_factory(void);
+void ensure_conf_dirs(void);
+
+// Preset Management
+void save_preset_file(const char *name);
+void load_preset_file(const char *name, bool quiet);
+void active_preset_name(char *out, size_t outsz);
+void set_active_preset(const char *name);
+void list_presets(char ***names, int *count);
+
+void up60p_options_from_settings(up60p_options *dst, const Settings *src);
+
+void settings_from_up60p_options(Settings *dst, const up60p_options *src);
 
 extern Settings S;
 extern Settings DEF;
