+++ myUpscaler/up60p_utils.c
@@ -1,6 +1,9 @@
 #include "up60p_utils.h"
+#include "up60p_common.h"
+#include "up60p_cli.h"
+#include ""
 
-
+up60p_log_callback global_log_cb = NULL;
 
 volatile sig_atomic_t cancel_requested = 0;
