+++ myUpscaler/upscaler/Up60PBridging.h
@@ -1,6 +1,6 @@
 // Up60PBridging.h
 #import "up60p.h"
-
-#ifdef UP60P_LIBRARY_MODE
-extern void (*global_log_cb)(const char *);
-#endif
+//
+//#ifdef UP60P_LIBRARY_MODE
+//extern void (*global_log_cb)(const char *);
+//#endif
