+++ myUpscaler/upscaler/Up60PBridging.m
@@ -1,6 +1,6 @@
 // Up60PBridging.m
 #import "Up60PBridging.h"
-
-#ifdef UP60P_LIBRARY_MODE
-void (*global_log_cb)(const char *) = NULL;
-#endif
+//
+//#ifdef UP60P_LIBRARY_MODE
+//void (*global_log_cb)(const char *) = NULL;
+//#endif
